requirements.txt
chartkick
django-admin-honeypot==1.1
django-cors-headers==3.0.2
django-debug-toolbar~=1.11
django-extensions<3
django-filter~=1.1
django-pdb
django-rest-framework
django-rosetta==0.9.3
django-tabbed-admin==1.0.4
django-taggit==0.24
django-test-without-migrations
django-widget-tweaks
Django<1.12
html2text
ipython
luhn
lxml
mercadopago
postgis
psycopg2
pyPdf
python-dateutil
python-memcached
reportlab
requests
setuptools
unicodecsv
yolk
zeep
